+++ packages/cashscript/src/libauth-template/utils.ts
@@ -1,6 +1,6 @@
 import { AbiFunction, AbiInput, Artifact, bytecodeToScript, formatBitAuthScript } from '@cashscript/utils';
 import { HashType, LibauthTokenDetails, SignatureAlgorithm, TokenDetails } from '../interfaces.js';
-import { hexToBin, binToHex, isHex, decodeCashAddress, type WalletTemplateScenarioBytecode } from '@bitauth/libauth';
+import { hexToBin, binToHex, isHex, decodeCashAddress, type WalletTemplateScenarioBytecode, Input, assertSuccess, decodeAuthenticationInstructions, AuthenticationInstructionPush } from '@bitauth/libauth';
 import { EncodedFunctionArgument } from '../Argument.js';
 import { zip } from '../utils.js';
 import SignatureTemplate from '../SignatureTemplate.js';
@@ -111,4 +111,16 @@ interface LibauthTemplateTokenDetails {
 
 export const lockingBytecodeIsSetToSlot = (lockingBytecode?: WalletTemplateScenarioBytecode | ['slot']): boolean => {
   return Array.isArray(lockingBytecode) && lockingBytecode.length === 1 && lockingBytecode[0] === 'slot';
+};
+
+export const getSignatureAndPubkeyFromP2PKHInput = (
+  libauthInput: Input,
+): { signature: Uint8Array; publicKey: Uint8Array } => {
+  const inputData = (assertSuccess(
+    decodeAuthenticationInstructions(libauthInput.unlockingBytecode))
+  ) as AuthenticationInstructionPush[];
+  const signature = inputData[0].data;
+  const publicKey = inputData[1].data;
+
+  return { signature, publicKey };
 };